student_assistant/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.75.0
httpx==0.27.2
faiss-cpu==1.8.0
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.9.2
langgraph==1.1.6
python-multipart==0.0.24
PyMuPDF==1.27.2.2
redis==5.2.1
